// frontend/api/zod/errorModelSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { errorDetailSchema } from "./errorDetailSchema.ts";
import { z } from "zod/v4";

export const errorModelSchema = z.object({
  $schema: z.optional(
    z.url().describe("A URL to the JSON Schema for this object."),
  ),
  detail: z.optional(
    z
      .string()
      .describe(
        "A human-readable explanation specific to this occurrence of the problem.",
      ),
  ),
  get errors() {
    return z
      .array(errorDetailSchema)
      .describe("Optional list of individual error details")
      .nullish();
  },
  instance: z.optional(
    z
      .url()
      .describe(
        "A URI reference that identifies the specific occurrence of the problem.",
      ),
  ),
  status: z.optional(z.int().describe("HTTP status code")),
  title: z.optional(
    z
      .string()
      .describe(
        "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
      ),
  ),
  type: z.optional(
    z
      .url()
      .default("about:blank")
      .describe(
        "A URI reference to human-readable documentation for the error.",
      ),
  ),
});